pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aleph-agent-mcp"
version = "0.1.0"
description = "MCP server for autonomous VM provisioning on Aleph Cloud"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastmcp>=2.0",
    "aleph-client>=1.9.0",
    "pydantic-settings>=2.0",
]

[project.scripts]
aleph-agent-mcp = "aleph_agent_mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aleph_agent_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
]
